Extract argument parsing, timer out of main
