+++ tooling/pipeline/src/services/sqlite/helpers/sqlite_connection_helpers.cc
